+++ designer/restmodel/src/main/scala/pl/touk/nussknacker/restmodel/PerformSingleExecutionResponse.scala
@@ -1,0 +1,6 @@
+package pl.touk.nussknacker.restmodel
+
+import io.circe.generic.JsonCodec
+
+@JsonCodec
+final case class PerformSingleExecutionResponse(isSuccess: Boolean, msg: String)